refactor: Criando o arquivo SugestaoCategoriaService
Essa nova classe é a nova responsável pelo metodo de sugerirCategoria do transacaoService, assim a refatoração já delegou duas funções que o transacaoService era responsável
